requirements.txt
aiogram==3.3.0
aiohttp==3.9.1
SQLAlchemy==2.0.25
alembic==1.13.1
pandas==2.1.4
numpy==1.26.3
python-binance==1.0.19
python-telegram-bot==20.7
beautifulsoup4==4.12.2
requests==2.31.0
aiosqlite==0.19.0
pytz==2023.3
python-dotenv==1.0.0
asyncio==3.4.3
ccxt==4.1.87
ta==0.10.2  # Technical Analysis library
plotly==5.18.0  # For charts
kaleido==0.2.1  # For static chart export
aiofiles==23.2.1
APScheduler==3.10.4
cryptography==41.0.7
web3==6.11.3  # For blockchain interactions
eth-account==0.10.0
eth-utils==2.3.1
tronpy==0.4.0  # For TRON blockchain
solana==0.30.2  # For Solana blockchain
python-dateutil==2.8.2
ujson==5.9.0
uvloop==0.19.0; sys_platform != "win32"  # Performance improvement for Unix systems 